Fix WERROR: ThreadMessageQueue2 size cast and ViewerCommand init
Cast queue size_type to int; initialize m_thumtoo in the member-init list
when HAVE_THUMTOO so -Weffc++ is satisfied.

+++ src/job/thread_message_queue2.hpp
@@ -50,7 +50,7 @@ public:
   int size() const
   {
     std::unique_lock<std::mutex> lock(m_mutex);
-    return m_queue.size();
+    return static_cast<int>(m_queue.size());
   }
 
   bool empty() const
